--- foldersTest.js
@@ -1,5 +1,0 @@
-const filesystem = require('./filesystem');
-
-filesystem.displayFolders('./filesystem').then((response) => {
-    console.log(response)
-});
