+++ test/allFilesTest.js
