+++ test/editorFileTest.js
